좋아요 취소시 전체 삭제 수정

+++ BackEnd/controllers/freeBoardsController.js
@@ -180,7 +180,7 @@ exports.thumbsUp = async (req, res) => {
 
     if (result) {
       //삭제
-      await FreeBoardLike.destroy({ where: { freeboard_id: post_id } });
+      await FreeBoardLike.destroy({ where: { freeboard_id: post_id, user_id :primaryKey } });
     } else {
       //추가
       await FreeBoardLike.create({
